requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sentence-transformers==2.2.2
transformers==4.30.2
huggingface_hub==0.14.1
pydantic==2.5.0
python-json-logger==2.0.4
faiss-cpu==1.7.4
numpy==1.24.3